+++ packages/carbon-graphs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed 
+  * Fixed retaining legend state in paired result graph during reflow.
+ 
 ## 2.17.0 - (April 27, 2021)
 
 * Changed